lab5/requirements.txt
numpy>=1.20.0
numba>=0.56.0
Pillow>=8.0.0
